Version 1.1 (fused)
Fused the server and original versions into a single script.  Improved
the handling of reindexing a genome to prevent potential collisions
between old data and new data.  This will new require the passing of
either —standAlone or —cluster on the command line in order to how
parallel jobs are handled.