Bug Fix in Indexing
Fixed a bug that could cause certain repeat regions to not be properly
indexed (due to change in the casing of the FASTA files).  In order to
ensure that these sites are properly indexed, you will need to rerun
your genome indexing with this new version (although you may not want
to for the reason below).  This will probably be the final version of
dsNickFury 1.  dsNickFury 2 is now written (the bug was found during
the writing of that version).  This new version has several major
performance improvements and switching is highly recommended.